Apply micro-optimization to atoms
The vast majority of atoms are made up of ASCII characters; atoms that
contain extended latin-1 characters (e.g., accented characters) are
extremely uncommon. This commit modifies the BertTerm::atom method to
avoid doing an ISO-8859-1 (latin-1) decode if all the bytes read are
within the ASCII range, i.e., smaller than 128.

The optimization has not been applied to BertTerm::atom_utf8 because
UTF-8 atoms are not yet very common.  If they become more common, or
the default, we'll revisit.