Interleave parsing and pretty printing
In previous versions of ppbert, the high-level flow for .bert2 files
was:

1. Read whole file in a Vec<u8>;
2. Parse the whole Vec<u8> into a Vec<BertTerm>;
3. Pretty print each BertTerm.

I was always uneasy about this flow, because it required more memory
to hold all the BertTerms in memory.  I also did not like the delay
between the moment the program started (reading and parsing) and the
moment it started printing if the input file was large.

With this commit, the parsing and pretty printing of .bert2 files and
disk_log files are interleaved.  (Technically, it's also true for
.bert files, but since they contain only one term, in practice it
changes nothing.)