Simplify error messages
The error messages were built with three methods:

1. the BertError::offset method returned the offset (if any) for the error;
2. the BertError::extra_info returned an optional String (allocation!)
   giving extra information for the error;
3. Display::fmt for BertError gave the basic error message and tacked
   on the extra parts.

Though there is a bit more repetition now and there's a chance that
the error messages will not be uniform, all the logic for printing out
an error is moved into Display::fmt. This should make it easier to add
new errors and it removes the unnecessary allocations.